Update icon test case
